Separate login functionality unit test
